% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AI_random_forest.R
\name{AI_randomForest}
\alias{AI_randomForest}
\title{Title}
\usage{
AI_randomForest(
  pheno_object = NULL,
  geno_omic_object = NULL,
  geno_omic_test_object = NULL,
  response = NULL,
  gen_name = NULL,
  core = NULL,
  message = TRUE,
  center = TRUE,
  para_tunning = FALSE,
  ranForest_paras_tunning = c(mtry_tune = NULL),
  ntree = 500,
  mtry = NULL,
  maxnodes = NULL,
  importance = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
