% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AI_RidgeRegression_Lasso.R
\name{AI_RidgeRegression_Lasso}
\alias{AI_RidgeRegression_Lasso}
\title{Title}
\usage{
AI_RidgeRegression_Lasso(
  pheno_object = NULL,
  geno_omic_object = NULL,
  geno_omic_test_object = NULL,
  response = NULL,
  gen_name = NULL,
  core = NULL,
  message = TRUE,
  scale = TRUE,
  para_tunning = FALSE,
  lasso_paras_tunning = c(lambda_tune = NULL),
  lambda = NULL,
  GS_model = c("Lasso", "Ridge_Regression"),
  ...
)
}
\arguments{
\item{GS_model}{}
}
\description{
Title
}
