% Generated package documentation placeholder; prefer roxygen source for future edits.
\name{AI_svm_cv}
\alias{AI_svm_cv}
\title{AI_svm_cv}
\usage{
AI_svm_cv(y, omics, tst, response_family = "gaussian", scaling = FALSE, 
    centering = TRUE, omic_count, svm_kernel = "Gaussian", sigma_value = NULL,
    C_value = 1, degree_value = 3, scale_value = 1, offset_value = 0, 
    gamma_value = NULL, svm_type = "eps-regression") 
}
\arguments{
\item{y}{Parameter passed to the y workflow option.}

\item{omics}{Parameter passed to the omics workflow option.}

\item{tst}{Parameter passed to the tst workflow option.}

\item{response_family}{Response family used by the model workflow.}

\item{scaling}{Logical indicating whether predictors should be scaled.}

\item{centering}{Logical indicating whether predictors should be centered.}

\item{omic_count}{Parameter passed to the omic_count workflow option.}

\item{svm_kernel}{Parameter passed to the svm_kernel workflow option.}

\item{sigma_value}{Parameter passed to the sigma_value workflow option.}

\item{C_value}{Parameter passed to the C_value workflow option.}

\item{degree_value}{Parameter passed to the degree_value workflow option.}

\item{scale_value}{Parameter passed to the scale_value workflow option.}

\item{offset_value}{Parameter passed to the offset_value workflow option.}

\item{gamma_value}{Parameter passed to the gamma_value workflow option.}

\item{svm_type}{Parameter passed to the svm_type workflow option.}

}
\description{
Documentation for \code{AI_svm_cv}. This topic is generated to keep exported package APIs documented and synchronized with the current function signature.
}
