% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AI_extreme_gradient_boosting.R
\name{AI_Xgb}
\alias{AI_Xgb}
\title{Title Extreme Gradient Boosting Machine Learning Genomic Selection Pipeline
Hyper-parameter tunning  of the parameters is allowed if desired by user
The parameters are:
Iter_tune : number of boosting iterations
learnining_rate_tune:  simply means how fast the model learns.
Each tree added modifies the overall model.
The magnitude of the modification is controlled by learning rate.
The lower the learning rate, the slower the model learns.
The advantage of slower learning rate is that the model becomes more robust
and efficient.}
\usage{
AI_Xgb(
  pheno_object = NULL,
  geno_omic_object = NULL,
  geno_omic_test_object = NULL,
  response = NULL,
  gen_name = NULL,
  message = TRUE,
  center = TRUE,
  para_tunning = FALSE,
  xgb_paras_tunning = c(Iter_tune = NULL, learning_rate_tune = NULL, L2_tune = NULL,
    L1_tune = NULL),
  resample_method_tune = "cv",
  number_of_fold_tune = 5,
  learning_rate = 0.001,
  max_depth = 6,
  subsample = 0.5,
  booster = "gblinear",
  iteration = 5000,
  N_feature_impo = 10,
  core = NULL,
  ...
)
}
\arguments{
\item{pheno_object}{phenotypic object NA is allowed}

\item{geno_omic_object}{multi-omic data, NA not allowed}

\item{geno_omic_test_object}{multi-omic data for testing set if not present in geno_omic_object}

\item{response}{y variables/lables}

\item{gen_name}{column name containing individuals/genotypes}

\item{center}{standardization of the x-variables}

\item{para_tunning}{if user required parameter tunning}

\item{xgb_paras_tunning}{parameter to for tunning}

\item{learning_rate}{how slow/fast the model learn}

\item{max_depth}{max_depth refers to the number of leaves of each tree}

\item{subsample}{This help to reduce the correlation between results from individual learners.}

\item{booster}{to determine if the model is for regression or classification problem}

\item{iteration}{Number of iteration}

\item{N_feature_impo}{number of feature/ x variables to extract based on the importance/weight}

\item{core}{number of ram for paralllel job}

\item{...}{}
}
\description{
Title Extreme Gradient Boosting Machine Learning Genomic Selection Pipeline
Hyper-parameter tunning  of the parameters is allowed if desired by user
The parameters are:
Iter_tune : number of boosting iterations
learnining_rate_tune:  simply means how fast the model learns.
Each tree added modifies the overall model.
The magnitude of the modification is controlled by learning rate.
The lower the learning rate, the slower the model learns.
The advantage of slower learning rate is that the model becomes more robust
and efficient.
}
