% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asreml_herit_varCov_output.R
\name{asreml_herit_varCov}
\alias{asreml_herit_varCov}
\title{Title}
\usage{
asreml_herit_varCov(
  model = NULL,
  heter_groups = NULL,
  VarCov_str = NULL,
  heter_resid = FALSE,
  G_list = NULL,
  Inter_Gen_pos = NULL,
  Gen_pos = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
