% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asreml_engine_output.R
\name{asreml_mod_outputOLD}
\alias{asreml_mod_outputOLD}
\title{Title}
\usage{
asreml_mod_outputOLD(
  mod_asreml = NULL,
  pheno_data = NULL,
  gkernel = NULL,
  gmatrix = NULL,
  omic1_kernel = NULL,
  omic2_kernel = NULL,
  omic3_kernel = NULL,
  heter_groups = NULL,
  gen_name = NULL,
  var_cov_str = NULL,
  heter_resid = NULL,
  pworkspace = 1e+15,
  maxit = 50,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
