% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asreml_variance_component_output.R
\name{asreml_varcomp_table}
\alias{asreml_varcomp_table}
\title{Extract an ASReml variance component table}
\usage{
asreml_varcomp_table(model_or_summary)
}
\arguments{
\item{model_or_summary}{ASReml model object or \code{summary.asreml} object.}
}
\value{
A data frame of ASReml variance components.
}
\description{
Extract an ASReml variance component table
}
\keyword{internal}
