% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_mod_execute_crossval.R
\name{bayes_mod_execute_crossval}
\alias{bayes_mod_execute_crossval}
\title{Title}
\usage{
bayes_mod_execute_crossval(
  pheno_object = NULL,
  response = NULL,
  weights = NULL,
  ETA = NULL,
  heter_groups = NULL,
  bayes_para = NULL,
  verbose = FALSE,
  core = NULL,
  test_set_val = NULL,
  cross_validation = NULL,
  eval_metrics = c("Accuracy", "Mean_Squared_Error", "Bias", "Root_Mean_Squared_Error",
    "Relative_Squared_Error", "Mean_Absolute_Error", "Mean_Absolute_Percent_Error"),
  ...
)
}
\arguments{
\item{eval_metrics}{}
}
\description{
Title
}
