% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesians_preprocess.R
\name{bayes_parameter_check}
\alias{bayes_parameter_check}
\title{Title}
\usage{
bayes_parameter_check(nIter = NULL, burnIn = NULL, thin = NULL, message = TRUE)
}
\arguments{
\item{nIter}{the number of iteration for MCMC chain}

\item{burnIn}{the number of burning}

\item{thin}{the number of thinning}

\item{message}{}
}
\description{
Title
}
\examples{
Title

}
