% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes_variance_components_organized.R
\name{Bayes_variance_components}
\alias{Bayes_variance_components}
\title{Title}
\usage{
Bayes_variance_components(
  Var_U_1 = NULL,
  Var_U_2 = NULL,
  Var_U_3 = NULL,
  Var_U_4 = NULL,
  Var_U = NULL,
  Var_E = NULL,
  genomic_h2 = NULL,
  Var_U_1_Se = NULL,
  Var_U_2_Se = NULL,
  Var_U_3_Se = NULL,
  Var_U_4_Se = NULL,
  Var_U_Se = NULL,
  Var_E_Se = NULL,
  genomic_h2_Se = NULL
)
}
\arguments{
\item{Var_E_Se}{}
}
\description{
Title
}
