% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blending_statt.R
\name{blending_stat}
\alias{blending_stat}
\title{Title}
\usage{
blending_stat(
  grm_kernel_data = NULL,
  pedigree_matrix = NULL,
  blending = TRUE,
  blending_value = 0.02
)
}
\arguments{
\item{blending_value}{}
}
\description{
Title
}
