% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blending_stat.R
\name{blending_statOLD}
\alias{blending_statOLD}
\title{Title}
\usage{
blending_statOLD(
  grm_kernel_data = NULL,
  pedigree_matrix = NULL,
  blending = TRUE,
  blending_value = 0.02
)
}
\arguments{
\item{blending_value}{}
}
\description{
Inference from ASRgenomics and modified to the suitability of objective in this package
}
