% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cals_Coeff_EBV_PEV_Rel_RHKS_glub.R
\name{Cal_Coeff_EBV_PEV_Rel_RHKS_glub}
\alias{Cal_Coeff_EBV_PEV_Rel_RHKS_glub}
\title{Title}
\usage{
Cal_Coeff_EBV_PEV_Rel_RHKS_glub(
  mod = NULL,
  gmatrix = NULL,
  gen_name = NULL,
  var_u = NULL,
  var_E = NULL,
  hetero = NULL,
  heter_groups = NULL,
  ...
)
}
\arguments{
\item{var_u}{genetic variance}

\item{...}{}

\item{beta}{the posterior samples of the random effects (marker/omic)}

\item{x_variable}{marker/omic data}
}
\description{
### This function calculate the Coefficient, EBV, PEV, SE, Reliability
}
