% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cals_Coeff_EBV_PEV_Rel_SE.R
\name{Cal_Coeff_EBV_PEV_Rel_SE}
\alias{Cal_Coeff_EBV_PEV_Rel_SE}
\title{Title}
\usage{
Cal_Coeff_EBV_PEV_Rel_SE(
  beta,
  x_variable,
  gen_name,
  var_u,
  hetero = NULL,
  heter_groups = NULL,
  gid_name = NULL,
  GS_model = NULL,
  ...
)
}
\arguments{
\item{beta}{the posterior samples of the random effects (marker/omic)}

\item{x_variable}{marker/omic data}

\item{var_u}{genetic variance}

\item{...}{}
}
\description{
### This function calculate the Coefficient, EBV, PEV, SE, Reliability
}
