% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_Recode_and_QC.R
\name{check_hapmap_columns}
\alias{check_hapmap_columns}
\title{Check HapMap Columns}
\usage{
check_hapmap_columns(data)
}
\arguments{
\item{data}{A data frame containing HapMap data.}
}
\value{
Throws an error if the columns do not match; otherwise, returns invisibly.
}
\description{
Checks and ensures that the first 11 columns of the HapMap data match the required format.
}
\examples{
hapmap <- data.frame(`rs#` = 1:3, alleles = c("A/T", "C/G", "A/A"), chrom = 1:3, pos = 1:3, strand = "+", assembly = NA,
                     center = NA, protLSID = NA, assayLSID = NA, panelLSID = NA, QCcode = NA, X1 = c("AA", "CC", "AT"), X2 = c("TA", "GC", "AA"))
names(hapmap)[1] <- "rs#"
PredictProR:::check_hapmap_columns(hapmap)

}
