% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BP_CM__CM_BP.R
\name{convert_BP_to_cM}
\alias{convert_BP_to_cM}
\title{Convert from basepairs to centimorgan}
\usage{
convert_BP_to_cM(pos_BP)
}
\arguments{
\item{pos_BP}{Numeric.  The position in basepairs.}
}
\value{
pos_CM The postion in centiMorgans
}
\description{
Convert from basepairs to centimorgan
}
\keyword{internal}
