% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BP_CM__CM_BP.R
\name{convert_CM_to_BP}
\alias{convert_CM_to_BP}
\title{Convert from centiMorgan to basepairs}
\usage{
convert_CM_to_BP(pos_CM)
}
\arguments{
\item{pos_CM}{Numeric.  The position in centiMorgan.}
}
\value{
pos_BP The postion in basepairs
}
\description{
Convert from centiMorgan to basepairs
}
\keyword{internal}
