% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_execute_crossval_para.R
\name{CrossVal}
\alias{CrossVal}
\title{Title}
\usage{
CrossVal(
  pheno_data = NULL,
  pheno_data_train = NULL,
  pheno_data_test = NULL,
  geno_data = NULL,
  omic1_data = NULL,
  omic2_data = NULL,
  omic3_data = NULL,
  gmatrix = NULL,
  gkernel = NULL,
  omic1_kernel = NULL,
  omic2_kernel = NULL,
  omic3_kernel = NULL,
  train_geno_data = NULL,
  train_omic1_data = NULL,
  train_omic2_data = NULL,
  train_omic3_data = NULL,
  test_geno_data = NULL,
  test_omic1_data = NULL,
  test_omic2_data = NULL,
  test_omic3_data = NULL,
  train_set = NULL,
  test_set = NULL,
  gmatrix_method = NULL,
  kernel_method = NULL,
  response = NULL,
  gen_name = NULL,
  cova = NULL,
  fixed = NULL,
  random = NULL,
  heter_resid = FALSE,
  heter_groups = NULL,
  var_cov_str = NULL,
  weights = NULL,
  nIter = NULL,
  burnIn = NULL,
  thin = NULL,
  GS_model = NULL,
  eval_metrics = c("Accuracy", "Mean_Squared_Error", "Bias", "Root_Mean_Squared_Error",
    "Relative_Squared_Error", "Mean_Absolute_Error", "Mean_Absolute_Percent_Error"),
  fixed_term_model_bayesian = "FIXED",
  rand_term_model_bayesian = NULL,
  cross_validation = c("Hold_Out", "Stratified_Hold_Out", "Repeated_Hold_Out",
    "Repeated_Stratified_Hold_Out", "K-Folds", "Stratified_K-Folds", "Repeated_K-Folds",
    "Repeated_Stratified_K-Folds", "Leave_one_Out", "CV1", "Repeated_CV1", "CV2",
    "Repeated_CV2"),
  test_size = 0.5,
  random_state = NULL,
  replication = 1,
  nFolds = NULL,
  CV = NULL,
  core = NULL,
  message = TRUE,
  center = TRUE,
  ...
)
}
\arguments{
\item{center}{}
}
\description{
Title
}
