% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV1_CV2_crossvalidation.R
\name{CV1_CV2}
\alias{CV1_CV2}
\title{Title}
\usage{
CV1_CV2(
  pheno_data,
  gen_name,
  heter_groups,
  CV = NULL,
  nFolds = NULL,
  random_state = NULL,
  replication = 1
)
}
\arguments{
\item{nFolds}{}
}
\description{
Title
}
