% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AI_deep_learning.R
\name{deep_learning_model}
\alias{deep_learning_model}
\title{Title}
\usage{
deep_learning_model(
  pheno_object = NULL,
  geno_omic_object = NULL,
  geno_omic_test_object = NULL,
  response = NULL,
  gen_name = NULL,
  message = TRUE,
  scale = TRUE,
  num_hidden_layers = 1,
  neurons_per_layer = NULL,
  learning_rate = 0.001,
  epochs = 10,
  batch_size = 32,
  para_tunning = FALSE,
  param_grid = NULL,
  validation_split = 0.2,
  early_stop = TRUE
)
}
\arguments{
\item{batch_size}{}
}
\description{
Title
}
