% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_Recode_and_QC.R
\name{double_code_to_IUPAC}
\alias{double_code_to_IUPAC}
\title{Convert Double Nucleotide Code to IUPAC Single-Letter Code}
\usage{
double_code_to_IUPAC(double_code)
}
\arguments{
\item{double_code}{A string representing a double nucleotide code (e.g., "AA", "AT").}
}
\value{
A single-letter IUPAC code corresponding to the double nucleotide code.
}
\description{
Converts a double nucleotide code to its corresponding IUPAC single-letter code.
}
\examples{
PredictProR:::double_code_to_IUPAC("AA") # Returns "A"
PredictProR:::double_code_to_IUPAC("AT") # Returns "W"

}
