% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{ellipsis_subset}
\alias{ellipsis_subset}
\title{Subset ellipsis based on a functions formals.}
\usage{
ellipsis_subset(ellipsis = NULL, fun = NULL)
}
\arguments{
\item{ellipsis}{The ellipsis (i.e., ...) (default = \code{NULL}).}

\item{fun}{The function to check formals and collect respective input from ... (default = \code{NULL}).}
}
\value{
A list with the ... corresponding to the function.
}
\description{
Sometimes we want the ellipsis to bear the input for more than one internal function.
Therefore we can split it according to the formals of each function.
}
\keyword{internal}
