% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETA_compiler.R
\name{ETA_compiler_bayes}
\alias{ETA_compiler_bayes}
\title{Title
This function does this following:
1. Extract all the random terms from the formula
2. Compile all the random terms and set the model for each term}
\usage{
ETA_compiler_bayes(
  fixed = NULL,
  random = NULL,
  GS_model = NULL,
  fixed_term_model_bayesian = NULL,
  rand_term_model_bayesian = NULL,
  pheno_data = NULL,
  geno_data = NULL,
  omic1_data = NULL,
  omic2_data = NULL,
  omic3_data = NULL,
  gen_name = NULL,
  ...
)
}
\arguments{
\item{fixed}{fixed effect. it must be in formula separated by +}

\item{random}{random effect, it must be formula separated by +}

\item{GS_model}{model defined by user}

\item{fixed_term_model_bayesian}{Model for fixed effect/term(s). default is FIXED}

\item{rand_term_model_bayesian}{Model for all random term(s) defined by the user}

\item{pheno_data}{phenotypic data. Dataframe or matrix is acceptable}

\item{geno_data}{snp/marker data. Dataframe or matrix is acceptable}

\item{omic1_data}{omic data .Dataframe or matrix is acceptable}

\item{omic2_data}{omic data .Dataframe or matrix is acceptable}

\item{omic3_data}{omic data .Dataframe or matrix is acceptable}

\item{gen_name}{Column name containing the genotype or individuals in the phenotypic data}

\item{...}{}
}
\description{
Title
This function does this following:
1. Extract all the random terms from the formula
2. Compile all the random terms and set the model for each term
}
