% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETA_compiler_bayes_GBLUP.R
\name{ETA_compiler_bayes_GBLUP}
\alias{ETA_compiler_bayes_GBLUP}
\title{Title}
\usage{
ETA_compiler_bayes_GBLUP(
  fixed = NULL,
  random = NULL,
  GS_model = NULL,
  fixed_term_model_bayesian = NULL,
  rand_term_model_bayesian = NULL,
  pheno_data = NULL,
  gmatrix = NULL,
  gkernel = NULL,
  omic1_kernel = NULL,
  omic2_kernel = NULL,
  omic3_kernel = NULL,
  gen_name = NULL,
  heter_groups = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
