% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_impo_xgb.R
\name{feature_impo_xgb}
\alias{feature_impo_xgb}
\title{Title}
\usage{
feature_impo_xgb(xgb_fit = NULL, X_train = NULL, N_feature_impo = NULL, ...)
}
\arguments{
\item{xgb_fit}{trained model}

\item{X_train}{genomic/omics data}

\item{N_feature_impo}{number of feature/ x variables to extract based on the importance/weight}
}
\description{
Title
}
