% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesians_preprocess.R
\name{fixed_term_model}
\alias{fixed_term_model}
\title{Title}
\usage{
fixed_term_model(
  fixed_term = NULL,
  fixed_term_model_bayesian = NULL,
  message = TRUE
)
}
\arguments{
\item{fixed_term}{all the fixed term, which is an output from fixed_terms function}

\item{fixed_term_model_bayesian}{default is FIXED}
}
\description{
Title
}
