% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_precheckk.R
\name{geno_precheck}
\alias{geno_precheck}
\title{Title}
\usage{
geno_precheck(
  object_geno = NULL,
  maf = 0.05,
  freqHet = 0.3,
  call_rate = 0.5,
  impute = TRUE,
  qc_filtering = TRUE,
  message = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
