% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_precheck.R
\name{geno_precheckOLD}
\alias{geno_precheckOLD}
\title{Title}
\usage{
geno_precheckOLD(
  object_geno = NULL,
  maf = 0.05,
  freqHet = 0.3,
  call_rate = 0.5,
  impute = TRUE,
  qc_filtering = TRUE,
  message = TRUE,
  ...
)
}
\arguments{
\item{object_geno}{snp/marker data. NA is allowed}

\item{maf}{minor allele frequency user defined It must be in percentage eg 0.05}

\item{freqHet}{heterozygosity accepted by the user. It must be in percentage eg 0.2}

\item{call_rate}{missing value defined by the user. It must be in percentage eg 0.4}

\item{impute}{If user want to impute missing value. Logical TRUE or FALSE}

\item{qc_filtering}{if user want to do quality control. Logical TRUE or FALSE}

\item{message}{If user want to print out message}
}
\description{
Though it is expected the user provide clean data.
To keep everything tidy for model fitting the function perform the
following objectives:
1. Check if the object_geno (geno/omic data) file contain records
2. Check if the object_geno is a matrix and if not we fix it for the user
3. Check the row and column names
4. Check that the marker/column are all in numeric and not character
5. Check if the allele dosages is not in 0, 1, 2 and it not, we fix it
6. Check for monomorphic marker and remove it.
7. Check for NA and remove it as defined by the user in the call_rate.
8. Impute missing value if user defined it
9. It check for MAF and Heterozygosity '
10. metadata is generated for qc and filtering.
   Thought it is actually expected to be a clean data.
11. The output is clean snp/marker data with NA,monomorphic, maf, heterozgous
    removed user defined some threshould value. If not default values are used.
   The output is assign attribute "pass" and declared class 'geno_data'
}
