% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETA_compiler.R
\name{gp_bayes_normalize_fixed_argument}
\alias{gp_bayes_normalize_fixed_argument}
\title{Compile ETA for Fixed Terms in Bayesian Models}
\usage{
gp_bayes_normalize_fixed_argument(
  fixed = NULL,
  fixed_term_model_bayesian = NULL
)
}
\arguments{
\item{fixed}{A formula specifying the fixed effects to be included in the model.}

\item{fixed_term_model_bayesian}{Optionally specify the model for the fixed terms in the Bayesian framework.
This is usually "FIXED" but can be left NULL for automatic handling.}
}
\value{
A list of lists where each inner list represents an ETA component for a fixed term.
        Each component contains the model matrix (`X`) and the model type (`model`) for that fixed term.
}
\description{
This function prepares the ETA components for fixed terms in Bayesian genomic prediction models.
It uses the specified fixed effects from the phenotypic data and compiles them into a format suitable
for inclusion in a Bayesian analysis. This is particularly useful for preparing data for models
such as GBLUP, RKHS, and BRR.
}
\examples{
\dontrun{
# Assuming pheno_data is your phenotypic dataset and fixed_effect is your fixed effect formula:
result <- ETA_compiler_fixed_term(fixed = ~fixed_effect1 + fixed_effect2,
                                  pheno_data = pheno_data)
}
}
