% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistent_met_summaries.R
\name{gp_empirical_env_correlation}
\alias{gp_empirical_env_correlation}
\title{Compute empirical environment correlation matrix from per-env predictions.}
\usage{
gp_empirical_env_correlation(preds, gid_col, env_col, value_col)
}
\description{
Compute empirical environment correlation matrix from per-env predictions.
}
\keyword{internal}
