% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search_utilities_deep_learning.R
\name{grid_search_dp}
\alias{grid_search_dp}
\title{Title}
\usage{
grid_search_dp(
  X_train,
  y_train,
  param_grid,
  epochs,
  batch_size,
  validation_split = 0.2,
  early_stop = TRUE
)
}
\arguments{
\item{X_train}{}

\item{y_train}{}

\item{param_grid}{}

\item{epochs}{}

\item{batch_size}{}

\item{validation_split}{}

\item{early_stop}{}
}
\description{
Title
}
