% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grm_kernel_precheckk.R
\name{grm_kernel_precheck}
\alias{grm_kernel_precheck}
\title{Title}
\usage{
grm_kernel_precheck(
  grm_kernel_data = NULL,
  pedigree_matrix = NULL,
  bending = TRUE,
  bend_value = 0.01,
  blending = FALSE,
  blending_value = 0.02,
  high_diag_cut_off = 1.2,
  low_diag_cut_off = 0.8,
  duplicate_cut_off = 0.95,
  rcn_cutoff = 1e-12,
  optimize_diagonal = FALSE,
  optimize_duplicate = FALSE,
  message = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
