% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_Recode_and_QC.R
\name{hmp_qc_recode_legacy}
\alias{hmp_qc_recode_legacy}
\title{Quality Control and Recoding for HapMap Data}
\usage{
hmp_qc_recode_legacy(
  hapmap_file_name = NULL,
  hapmap_file_path = NULL,
  maf_threshold = 0.01,
  het_threshold = 0.1,
  ind_call_rate_threshold = 0.9,
  snp_call_rate_threshold = 0.9,
  impute = TRUE,
  recode_format = "0,1,2",
  out_put_map = TRUE,
  message = TRUE,
  ...
)
}
\arguments{
\item{hapmap_file_name}{A string specifying the name of the HapMap file to be processed. If NULL, `hapmap` must be provided.}

\item{hapmap_file_path}{A string specifying the path to the directory containing the HapMap file. If NULL, `hapmap` must be provided.}

\item{maf_threshold}{A numeric value specifying the threshold for minor allele frequency (MAF). Markers below this threshold will be removed.}

\item{het_threshold}{A numeric value specifying the threshold for heterozygosity. Markers above this threshold will be removed.}

\item{ind_call_rate_threshold}{A numeric value specifying the threshold for individual call rate. Individuals below this threshold will be removed.}

\item{snp_call_rate_threshold}{A numeric value specifying the threshold for SNP call rate. SNPs below this threshold will be removed.}

\item{impute}{A logical value indicating whether missing data should be imputed. If TRUE, missing values will be imputed.}

\item{recode_format}{A string specifying the format for recoding genotype data. Can be "0,1,2" or "-1,0,1".}

\item{out_put_map}{A logical value indicating whether to output a map of SNP information.}

\item{message}{A logical value indicating whether to display messages during processing.}

\item{...}{Additional arguments to be passed to underlying functions.}
}
\value{
A list containing three elements:
  \item{snps_matrix}{A matrix or data.table of the processed and possibly recoded SNP data.}
  \item{snp_map}{An optional data.table containing SNP mapping information, depending on `out_put_map`.}
  \item{qc_metrics_and_summary_stat}{A data.frame summarizing the quality control metrics and actions taken.}
}
\description{
Performs quality control checks and recodes HapMap genotype data. It allows for filtering based on minor allele frequency (MAF), heterozygosity, individual and SNP call rates, and optionally imputes missing data. The function can also recode genotype data into numeric formats.
}
\examples{
\dontrun{
  # Save the example data to a file
  example_hapmap <- data.table::data.table(
    `rs#` = c("rs1", "rs2", "rs3", "rs4"),
    alleles = c("A/T", "C/G", "A/A", "G/T"),
    chrom = c(1, 1, 2, 2),
    pos = c(100, 200, 300, 400),
    strand = c("+", "+", "-", "-"),
    assembly = c("v1", "v1", "v1", "v1"),
    center = c("C1", "C1", "C1", "C1"),
    protLSID = c("p1", "p1", "p1", "p1"),
    assayLSID = c("a1", "a1", "a1", "a1"),
    panelLSID = c("panel1", "panel1", "panel1", "panel1"),
    QCcode = c(NA, NA, NA, NA),
    Sample1 = c("AA", "CC", "AA", "GG"),
    Sample2 = c("TT", "GG", "AA", "TT"),
    Sample3 = c("AT", "CG", "AA", "GT")
  )
  data.table::fwrite(example_hapmap, "example_hapmap.txt", sep = "\t")

  # Run the function with the example data
  result <- hmp_qc_recode(hapmap_file_name = "example_hapmap.txt",
                          hapmap_file_path = ".",
                          maf_threshold = 0.05,
                          het_threshold = 0.1,
                          ind_call_rate_threshold = 0.95,
                          snp_call_rate_threshold = 0.95,
                          impute = TRUE,
                          recode_format = "0,1,2",
                          out_put_map = TRUE,
                          message = TRUE)
}

}
\keyword{internal}
