% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_Recode_and_QC.R
\name{IUPAC_hapmap_compatible}
\alias{IUPAC_hapmap_compatible}
\title{Standardize SNP Format in HapMap Data Frame}
\usage{
IUPAC_hapmap_compatible(hapmap)
}
\arguments{
\item{hapmap}{A data frame containing HapMap genotype data starting from the 12th column.}
}
\value{
A HapMap data frame with standardized SNP format.
}
\description{
Standardizes SNP format in a HapMap data frame to be IUPAC compatible.
}
\examples{
hapmap <- data.frame(
  rs = "m1", alleles = "A/T", chrom = 1, pos = 10, strand = "+",
  assembly = NA, center = NA, protLSID = NA, assayLSID = NA,
  panelLSID = NA, QCcode = NA, X1 = "AA", X2 = "TA"
)
IUPAC_hapmap_compatible(hapmap)

}
