% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AI_data_processing.R
\name{ML_data_processing}
\alias{ML_data_processing}
\title{Title}
\usage{
ML_data_processing(
  pheno_clean = NULL,
  response = NULL,
  geno_clean = NULL,
  gen_name = NULL,
  omic_clean = list()
)
}
\arguments{
\item{omic_clean}{}
}
\description{
Title
}
