% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_output_BayesA_B_C_BL_BRR.R
\name{mod_output_bayes}
\alias{mod_output_bayes}
\title{Title}
\usage{
mod_output_bayes(
  mod = NULL,
  ETA = NULL,
  gen_name = NULL,
  geno_data = NULL,
  omic1_data = NULL,
  omic2_data = NULL,
  omic3_data = NULL,
  omics_data_label = list(omic1_data = NULL, omic2_data = NULL, omic3_data = NULL),
  bayes_para = NULL,
  GS_model = NULL,
  ...
)
}
\description{
Title
}
