% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_output_bayes_RKHS.R
\name{mod_output_bayes_RKHS}
\alias{mod_output_bayes_RKHS}
\title{Title}
\usage{
mod_output_bayes_RKHS(
  mod = NULL,
  ETA = NULL,
  gen_name = NULL,
  gkernel = NULL,
  gmatrix = NULL,
  omic1_kernel = NULL,
  omic2_kernel = NULL,
  omic3_kernel = NULL,
  omics_kernel_label = list(omic1_kernel = NULL, omic2_kernel = NULL, omic3_kernel =
    NULL),
  pheno_data = NULL,
  heter_groups = NULL,
  bayes_para = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
