% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_geno_match.R
\name{pheno_geno_match}
\alias{pheno_geno_match}
\title{Title
This function perform the following task:}
\usage{
pheno_geno_match(
  object_pheno = NULL,
  object_geno = NULL,
  gen_name = NULL,
  test_set = NULL,
  train_set = NULL,
  message = TRUE,
  ...
)
}
\arguments{
\item{object_pheno}{clean phenotypic data}

\item{object_geno}{clean snp/omic/grm/kernel data}

\item{gen_name}{column name of the genotypes in the phenotypic data}

\item{...}{}
}
\description{
1. This check for NA, match the grm/kernel and geno/omic (M_matrix) to pheno_data
2. It also order the M_matrix, grm/kernel to the GID in pheno_data
3. Ordering is not important in asreml but important in Baysiand and machine learning models.
}
\details{
The output is a list with one or two element(s):
1. the clean M-matrix or grm file that perfectly match with the phenotypic records
2. test_set that might emmanate from the object_geno


 NOTE FOR Giovanni ASRgenomics only works with grm/kernel not with M_matrix that was why it not used
}
