% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_to_model.R
\name{phenotype_to_model}
\alias{phenotype_to_model}
\title{Title
Overall, this serve as gateway between phenotype-precheck function and readiness of
the phenotypic data for model fitting
The objective of this function is to do the following:
1. Check the output from phenotype-precheck function for pheno_data before declaring it for model fit.
   NA is allowed for response variable with the assumption that those individuals are the testing set.
2. Check output from phenotype-precheck function for pheno_data_train and
   check for NA. If NA present it will not be declare for model fit because
   NA is not expected in training set. If all good it will be declared for model fit
3. Check the output from phenotype-precheck function for pheno_data_test before declaring it for model fit
   NA is allowed for response variable here.
4. If pheno_data_train and pheno_data_test were present and pass through the pre-check process,
   it will be processed that is.
   1) it check that column name for both data match/the same
   2) Combined the dataset for model fit and prediction.
   It is assumed here that pheno_data is missing/not provided by the user.
5. The output will be pheno_data declared for model fit and test_set if pheno_data_test was provided.}
\usage{
phenotype_to_model(
  pheno_data = NULL,
  pheno_data_train = NULL,
  pheno_data_test = NULL,
  train_set = NULL,
  test_set = NULL,
  response = NULL,
  gen_name = NULL,
  ...
)
}
\arguments{
\item{pheno_data}{phenotypic object NA is allowed}

\item{pheno_data_train}{phenotypic object for the training set. NA is allowed.
This is expected if pheno_data was not provided}

\item{pheno_data_test}{phenotypic object for the testing set. NA is allowed}

\item{test_set}{}

\item{response}{y variables/lables}

\item{gen_name}{column name containing individuals/genotypes}
}
\description{
Title
Overall, this serve as gateway between phenotype-precheck function and readiness of
the phenotypic data for model fitting
The objective of this function is to do the following:
1. Check the output from phenotype-precheck function for pheno_data before declaring it for model fit.
   NA is allowed for response variable with the assumption that those individuals are the testing set.
2. Check output from phenotype-precheck function for pheno_data_train and
   check for NA. If NA present it will not be declare for model fit because
   NA is not expected in training set. If all good it will be declared for model fit
3. Check the output from phenotype-precheck function for pheno_data_test before declaring it for model fit
   NA is allowed for response variable here.
4. If pheno_data_train and pheno_data_test were present and pass through the pre-check process,
   it will be processed that is.
   1) it check that column name for both data match/the same
   2) Combined the dataset for model fit and prediction.
   It is assumed here that pheno_data is missing/not provided by the user.
5. The output will be pheno_data declared for model fit and test_set if pheno_data_test was provided.
}
