% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics_asreml.R,
%   R/summary_statistics_plot.R
\name{plot_acc}
\alias{plot_acc}
\title{Title}
\usage{
plot_acc(mod, response, ...)

plot_acc(mod, response, ...)
}
\arguments{
\item{...}{}
}
\description{
Title

Title
}
