% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics_AI.R
\name{plot_acc_AI}
\alias{plot_acc_AI}
\title{Title}
\usage{
plot_acc_AI(
  mod = NULL,
  pheno_object = NULL,
  response = NULL,
  test_set = NULL,
  GS_model = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
