% Generated package documentation placeholder; prefer roxygen source for future edits.
\name{process_batch}
\alias{process_batch}
\title{process_batch}
\usage{
process_batch(vcf_data, het_threshold, recode_format) 
}
\arguments{
\item{vcf_data}{Parameter passed to the vcf_data workflow option.}

\item{het_threshold}{Parameter passed to the het_threshold workflow option.}

\item{recode_format}{Parameter passed to the recode_format workflow option.}

}
\description{
Documentation for \code{process_batch}. This topic is generated to keep exported package APIs documented and synchronized with the current function signature.
}
