% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_geno_omic_kernel_gmatrix_data.R
\name{process_omic_data}
\alias{process_omic_data}
\title{Title}
\usage{
process_omic_data(
  omic_data = NULL,
  train_omic_data = NULL,
  test_omic_data = NULL,
  kernel_method = NULL,
  gen_name = NULL,
  test_set = NULL,
  train_set = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
