% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesians_preprocess.R
\name{random_term_model}
\alias{random_term_model}
\title{Validate and Adjust Model Specification for Random Terms}
\usage{
random_term_model(
  rand_terms = NULL,
  GS_model = NULL,
  gen_name = NULL,
  rand_terms_model_bayesian = NULL,
  message = TRUE
)
}
\arguments{
\item{rand_terms}{A character vector specifying the random effect terms to be included in the model.}

\item{GS_model}{A character string or vector specifying the general statistical model(s) to be applied,
such as "BRR", "BayesA", "BayesB", "BayesC", "BL", "RKHS".}

\item{gen_name}{A character string specifying the name of the genetic effect in the model.}

\item{rand_terms_model_bayesian}{A character vector specifying the model(s) for each random term.
If not specified or inconsistent, default adjustments are made based on `GS_model` and available terms.}

\item{message}{A logical indicating whether to display warning messages for missing or incorrect model specifications.}
}
\value{
A character vector with the adjusted or validated model specifications for each random term.
}
\description{
This function evaluates and adjusts the model specifications for random terms in a Bayesian analysis setting,
ensuring they align with the specified general and random term models. It handles various scenarios including
interactions, the presence of genetic names, and checks for consistency with available Bayesian models.
}
\details{
The function ensures that the model specifications for random terms are correctly set for Bayesian analysis.
It verifies the presence and appropriateness of the genetic effect name, handles interaction terms,
and adjusts model specifications based on the general model (`GS_model`) provided. If `GS_model` is specified
but `rand_terms_model_bayesian` is missing or incomplete, the function attempts to intelligently assign models
to each term, defaulting to "BRR" where necessary and using "RKHS" for multi-kernel interactions.
}
\examples{
rand_terms <- c("Genotype", "Genotype:Environment")
GS_model <- "BRR"
gen_name <- "Genotype"
rand_terms_model_bayesian <- c("BRR", "RKHS")
adjusted_models <- random_term_model(rand_terms = rand_terms,
                                     GS_model = GS_model,
                                     gen_name = gen_name,
                                     rand_terms_model_bayesian = rand_terms_model_bayesian,
                                     message = TRUE)
print(adjusted_models)

}
