% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_Recode_and_QC.R
\name{remove_multiallelic_markers}
\alias{remove_multiallelic_markers}
\title{Remove Multiallelic Markers}
\usage{
remove_multiallelic_markers(hapmap)
}
\arguments{
\item{hapmap}{A data frame containing HapMap genotype data.}
}
\value{
A HapMap data frame with multiallelic markers removed.
}
\description{
Removes markers with more than three alleles from a HapMap data frame.
}
\examples{
hapmap <- data.frame(rs = 1:3, alleles = c("A/T/G", "C/G", "A/A"), X1 = c("A", "C", "A"), X2 = c("T", "G", "A"))
PredictProR:::remove_multiallelic_markers(hapmap)

}
