% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_matrix.R
\name{sparse_matrix}
\alias{sparse_matrix}
\title{Title
Convert dense matrix to sparse form}
\usage{
sparse_matrix(grm_kernel_data = NULL, inverse = NULL, drop_zero = TRUE)
}
\arguments{
\item{drop_zero}{}
}
\description{
Title
Convert dense matrix to sparse form
}
