% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_Recode_and_QC.R
\name{standardize_snp_format}
\alias{standardize_snp_format}
\title{Standardize SNP Format}
\usage{
standardize_snp_format(snp)
}
\arguments{
\item{snp}{A string representing a SNP in either single or double format.}
}
\value{
A standardized SNP in IUPAC single-letter code.
}
\description{
Standardizes SNP format by converting double nucleotide codes to IUPAC single-letter codes.
}
\examples{
PredictProR:::standardize_snp_format("A/T") # Returns "W"
PredictProR:::standardize_snp_format("AA")  # Returns "A"

}
