% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics_asreml.R
\name{summary_statistics_asreml}
\alias{summary_statistics_asreml}
\title{Title}
\usage{
summary_statistics_asreml(
  mod = NULL,
  response = NULL,
  eval_metrics = c("Accuracy", "Mean_Squared_Error", "Bias", "Root_Mean_Squared_Error",
    "Relative_Squared_Error", "Mean_Absolute_Error", "Mean_Absolute_Percent_Error"),
  heter_groups = NULL,
  predicted_value = NULL,
  pred_heter_groups = NULL,
  variance_components = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
