% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics_plot.R
\name{summary_statistics_bayes}
\alias{summary_statistics_bayes}
\title{Title}
\usage{
summary_statistics_bayes(
  mod = NULL,
  eval_metrics = c("Accuracy", "Mean_Squared_Error", "Bias", "Root_Mean_Squared_Error",
    "Relative_Squared_Error", "Mean_Absolute_Error", "Mean_Absolute_Percent_Error"),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
