% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics_hybrid.R
\name{summary_statistics_hybrid}
\alias{summary_statistics_hybrid}
\title{Summarize Hybrid Prediction Outputs}
\usage{
summary_statistics_hybrid(
  predicted_object,
  response,
  female_parent,
  male_parent,
  mode,
  model_type
)
}
\arguments{
\item{predicted_object}{Hybrid prediction table returned by a hybrid model path.}

\item{response}{Response column name.}

\item{female_parent}{Female parent column name.}

\item{male_parent}{Male parent column name.}

\item{mode}{Hybrid mode label such as `hybrid_asreml`, `hybrid_bayes`,
`hybrid_gp`, `hybrid_ml`, or `hybrid_dl`.}

\item{model_type}{Model name.}
}
\value{
A list of hybrid summary data frames.
}
\description{
Summarize Hybrid Prediction Outputs
}
