% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdout_cv.R
\name{train_test_split}
\alias{train_test_split}
\title{Title}
\usage{
train_test_split(
  pheno_data = NULL,
  gen_name = NULL,
  response = NULL,
  test_size = 0.2,
  random_state = NULL,
  replication = 1,
  method = c("stratified", "unstratified"),
  message = TRUE,
  ...
)
}
\arguments{
\item{message}{}
}
\description{
Title
}
