% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AI_K-Nearest_Neighbors.R
\name{AI_knn}
\alias{AI_knn}
\title{Title}
\usage{
AI_knn(
  pheno_object = NULL,
  geno_omic_object = NULL,
  geno_omic_test_object = NULL,
  response = NULL,
  gen_name = NULL,
  core = NULL,
  message = TRUE,
  scale = TRUE,
  para_tunning = FALSE,
  knn_paras_tunning = c(k_tune = NULL),
  k = 5,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
