% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AI_Partial_least_square.R
\name{AI_pls}
\alias{AI_pls}
\title{Title}
\usage{
AI_pls(
  pheno_object = NULL,
  geno_omic_object = NULL,
  geno_omic_test_object = NULL,
  response = NULL,
  gen_name = NULL,
  message = TRUE,
  scale = TRUE,
  para_tunning = FALSE,
  ncomp = 3,
  pls_paras_tunning = c(ncomp = 10),
  resample_method_tune = "cv",
  N_feature_impo = 10,
  core = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
