% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AI_support_vector_machines.R
\name{AI_svm}
\alias{AI_svm}
\title{Title}
\usage{
AI_svm(
  pheno_object = NULL,
  geno_omic_object = NULL,
  geno_omic_test_object = NULL,
  response = NULL,
  gen_name = NULL,
  core = NULL,
  message = TRUE,
  center = TRUE,
  para_tunning = FALSE,
  svm_paras_tunning = c(c_tune = NULL),
  c = 1,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
