% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asreml_herit_compound_symmetry.R
\name{asreml_herit_CSM}
\alias{asreml_herit_CSM}
\title{Title}
\usage{
asreml_herit_CSM(
  model = mod,
  heter_groups = NULL,
  heter_resid = FALSE,
  G_list = NULL,
  Inter_Gen_pos = NULL,
  Gen_pos = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
