% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asreml_herit_varcov_outputt.R
\name{asreml_herit_varCov}
\alias{asreml_herit_varCov}
\title{Title}
\usage{
asreml_herit_varCov(
  model = NULL,
  heter_groups = NULL,
  var_cov_str = NULL,
  heter_resid = FALSE,
  G_list = NULL,
  inter_gen_pos = NULL,
  gen_pos = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
