% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asreml_utilisOLDD.R
\name{asreml_utilisOLDD}
\alias{asreml_utilisOLDD}
\title{Title}
\usage{
asreml_utilisOLDD(
  fixed = NULL,
  random = NULL,
  cova = NULL,
  GS_model = NULL,
  response = NULL,
  pheno_data = NULL,
  gmatrix = NULL,
  gkernel = NULL,
  omic1_kernel = NULL,
  omic2_kernel = NULL,
  omic3_kernel = NULL,
  gen_name = NULL,
  heter_groups = NULL,
  heter_resid = FALSE,
  var_cov_str = NULL,
  weights = NULL,
  core = NULL,
  workspace = 1e+08,
  maxit = 50,
  ...
)
}
\arguments{
\item{var_cov_str}{h@param weights}

\item{...}{}
}
\description{
Title
}
