% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_finalize_A_B_C_BL_BRR.R
\name{bayes_finalize_A_B_C_BL_BRR}
\alias{bayes_finalize_A_B_C_BL_BRR}
\title{Title}
\usage{
bayes_finalize_A_B_C_BL_BRR(
  fixed = NULL,
  random = NULL,
  GS_model = NULL,
  response = NULL,
  weights = NULL,
  fixed_term_model_bayesian = NULL,
  rand_term_model_bayesian = NULL,
  pheno_data = NULL,
  geno_data = NULL,
  omic1_data = NULL,
  omic2_data = NULL,
  omic3_data = NULL,
  gen_name = NULL,
  nIter = NULL,
  burnIn = NULL,
  thin = NULL,
  omics_data_label = NULL
)
}
\arguments{
\item{omics_data_label}{}
}
\description{
Title
}
