% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_finalize_RKHS_GBLUPBRR.R
\name{bayes_finalize_RKHS_GBLUPBRR}
\alias{bayes_finalize_RKHS_GBLUPBRR}
\title{Title}
\usage{
bayes_finalize_RKHS_GBLUPBRR(
  fixed = NULL,
  random = NULL,
  GS_model = NULL,
  response = NULL,
  weights = NULL,
  fixed_term_model_bayesian = NULL,
  rand_term_model_bayesian = NULL,
  pheno_data = NULL,
  gkernel = NULL,
  gmatrix = NULL,
  omic1_kernel = NULL,
  omic2_kernel = NULL,
  omic3_kernel = NULL,
  gen_name = NULL,
  nIter = NULL,
  burnIn = NULL,
  thin = NULL,
  heter_groups = NULL,
  omics_kernel_label = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
