% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_mod_execute.R
\name{bayes_mod_execute}
\alias{bayes_mod_execute}
\title{Title}
\usage{
bayes_mod_execute(
  pheno_data = NULL,
  response = NULL,
  weights = NULL,
  ETA = NULL,
  bayes_para = NULL,
  verbose = FALSE,
  core = NULL,
  ...
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
