% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_nFolds.R
\name{CV_nfolds}
\alias{CV_nfolds}
\title{Title}
\usage{
CV_nfolds(
  pheno_data = NULL,
  response = NULL,
  gen_name = NULL,
  nFolds = 5,
  random_state = NULL,
  method = c("stratified", "unstratified"),
  replication = 1
)
}
\arguments{
\item{method}{}
}
\description{
Title
}
