% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_learning_model_utility.R
\name{deep_learning_model_utility}
\alias{deep_learning_model_utility}
\title{Title}
\usage{
deep_learning_model_utility(
  X_train,
  y_train,
  num_hidden_layers,
  neurons_per_layer,
  learning_rate,
  epochs,
  batch_size
)
}
\arguments{
\item{batch_size}{}
}
\description{
Title
}
