% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_genomic_coding.R
\name{detect_genomic_coding}
\alias{detect_genomic_coding}
\title{Title
This function check genomic data for different coding
schemes: presence/absence or SNP data (coded as 0, 1, 2)}
\usage{
detect_genomic_coding(object_geno = NULL)
}
\arguments{
\item{object_geno}{}
}
\description{
Title
This function check genomic data for different coding
schemes: presence/absence or SNP data (coded as 0, 1, 2)
}
