% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETA_compiler.R
\name{ETA_compiler_fixed_term}
\alias{ETA_compiler_fixed_term}
\title{This function deal with fixed terms defined by the user and assign model to each fixed term}
\usage{
ETA_compiler_fixed_term(
  fixed = NULL,
  fixed_term_model_bayesian = NULL,
  pheno_data = NULL
)
}
\arguments{
\item{fixed}{fixed terms defined by the user}

\item{fixed_term_model_bayesian}{model for the fixed term and the default is FIXED}

\item{pheno_data}{phenotypic data}
}
\description{
This function deal with fixed terms defined by the user and assign model to each fixed term
}
