% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation_Metrics.R
\name{evaluation_metrics}
\alias{evaluation_metrics}
\title{Title}
\usage{
evaluation_metrics(
  y_observed = NULL,
  y_predicted = NULL,
  eval_metrics = c("Accuracy", "Mean_Squared_Error", "Bias", "Root_Mean_Squared_Error",
    "Relative_Squared_Error", "Mean_Absolute_Error", "Mean_Absolute_Percent_Error"),
  ...
)
}
\arguments{
\item{y_predicted}{}
}
\description{
Title
}
