% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesians_preprocess.R
\name{fixed_terms}
\alias{fixed_terms}
\title{Title}
\usage{
fixed_terms(fixed = NULL, pheno_data = NULL, ...)
}
\arguments{
\item{fixed}{fixed effect, only formula is accepted}

\item{pheno_data}{phenotypic data}

\item{...}{}
}
\description{
Title
}
\examples{
Title

}
