% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_precheckk.R
\name{geno_precheck}
\alias{geno_precheck}
\title{Title}
\usage{
geno_precheck(
  object_geno = NULL,
  qc_filtering = TRUE,
  maf_threshold = 0.05,
  het_threshold = 0.2,
  ind_call_rate_threshold = 0.9,
  snp_call_rate_threshold = 0.9,
  impute = TRUE,
  message = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
