% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_to_modell.R
\name{geno_to_model}
\alias{geno_to_model}
\title{Title
#' Title}
\usage{
geno_to_model(
  geno_data = NULL,
  train_geno_data = NULL,
  test_geno_data = NULL,
  map_data = NULL,
  message = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Overall, this serve as gateway between snp/marker-precheck function and readiness of
the snp/marker data for model fitting#'
The objective of this function is to do the following:
1. Check the output from geno-precheck function for geno_data before declaring it for model fit
2. If geno_data_train and geno_data_test were present and pass through the pre-check process,
  These will processed be processed that is:
   1) It check that column name (maker/snp) for both data match/the same
   2) Combined the dataset for model fit and prediction.
   It is assumed here that geno_data is missing/not provided by the user.
3. The output will be a matrix(geno_data) declared for model fit.
}
