% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grm_calculation.R
\name{grm_calculation}
\alias{grm_calculation}
\title{Title}
\usage{
grm_calculation(
  geno_clean = NULL,
  weight = NULL,
  method = c("VanRaden", "Weighted_VanRaden", "Yang", "Epistasis")
)
}
\arguments{
\item{method}{}
}
\description{
Title
}
