% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grm_kernel_diagnostic_fix.R
\name{grm_kernel_diagnostic_fix}
\alias{grm_kernel_diagnostic_fix}
\title{Title
This process was infer from ASRgenomics. It was modified and improved to suite the
objectives in this package}
\usage{
grm_kernel_diagnostic_fix(
  grm_kernel_data = NULL,
  high_diag_cut_off = 1.2,
  low_diag_cut_off = 0.8,
  duplicate_cut_off = 0.95,
  optimize_diagonal = FALSE,
  optimize_duplicate = FALSE
)
}
\arguments{
\item{message}{}
}
\description{
Title
This process was infer from ASRgenomics. It was modified and improved to suite the
objectives in this package
}
