% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grm_kernel_precheck.R
\name{grm_kernel_precheck}
\alias{grm_kernel_precheck}
\title{Title
grm_kernel_data is a genomic relationship matrix or relationship matrix calculated using
other omics data. The relationship matrix can be calculated using any method.
NA is not allowed in the relationship matrix
####
Checks
#######
1. It check if the matrix is square matrix/symmetry, if not we fix it for the user
2. It check if the matrix is positive definite, if not we fix it.
3. It check for NA. If present the engine will stop further analysis.}
\usage{
grm_kernel_precheck(
  grm_kernel_data = NULL,
  pedigree_matrix = NULL,
  bending = TRUE,
  bend_value = 0.01,
  blending = FALSE,
  blending_value = 0.02,
  high_diag_cut_off = 1.2,
  low_diag_cut_off = 0.8,
  duplicate_cut_off = 0.95,
  rcn_cutoff = 1e-12,
  optimize_diagonal = FALSE,
  optimize_duplicate = FALSE,
  message = TRUE,
  ...
)
}
\arguments{
\item{rcn_cutoff}{#the reciprocal conditional number of the inverse}
}
\description{
Title
grm_kernel_data is a genomic relationship matrix or relationship matrix calculated using
other omics data. The relationship matrix can be calculated using any method.
NA is not allowed in the relationship matrix
####
Checks
#######
1. It check if the matrix is square matrix/symmetry, if not we fix it for the user
2. It check if the matrix is positive definite, if not we fix it.
3. It check for NA. If present the engine will stop further analysis.
}
