% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_recode_and_qc.R
\name{hmp_qc_recode}
\alias{hmp_qc_recode}
\title{Title}
\usage{
hmp_qc_recode(
  hapmap_file_name = "HapMap_300_Aug25_Select_Chr_No_Filtering.hmp.txt",
  hapmap_file_path = "D:/PEA_BARI",
  hapmap = NULL,
  maf_threshold = 0.01,
  het_threshold = 0.1,
  ind_call_rate_threshold = 0.9,
  snp_call_rate_threshold = 0.9,
  hwe_threshold = 0.001,
  impute = FALSE,
  recode_format = "0,1,2",
  out_put_map = FALSE,
  message = TRUE,
  ...
)

hmp_qc_recode(
  hapmap_file_name = "HapMap_300_Aug25_Select_Chr_No_Filtering.hmp.txt",
  hapmap_file_path = "D:/PEA_BARI",
  hapmap = NULL,
  maf_threshold = 0.01,
  het_threshold = 0.1,
  ind_call_rate_threshold = 0.9,
  snp_call_rate_threshold = 0.9,
  hwe_threshold = 0.001,
  impute = FALSE,
  recode_format = "0,1,2",
  out_put_map = FALSE,
  message = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title

Title
}
