% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_calculation.R
\name{kernel_calculation}
\alias{kernel_calculation}
\title{Title
This function calculate the relationship for using different kernel methods
M_matrix_clean is not expected with NA value
The following kernel methods are available:
Gaussian, Linear, Exponential, Poly2, Poly3, Poly4. The number at the end of poly
means polynomial order 2, 3 and 4}
\usage{
kernel_calculation(
  M_matrix_clean = NULL,
  scale = TRUE,
  theta = NULL,
  alpha = 0.5,
  gamma = 1,
  smoothness_parameter = 1.5,
  length_scale = 2,
  method = NULL,
  message = TRUE,
  ...
)
}
\arguments{
\item{M_matrix_clean}{snp or omics matrix data.}

\item{...}{}
}
\description{
To DO: in the new version, option to impute missing data will be provided
}
