% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_loc_BayesA_B_C_BL_BRR.R
\name{M_matrix_bayes_mod_single_loc}
\alias{M_matrix_bayes_mod_single_loc}
\title{Title}
\usage{
M_matrix_bayes_mod_single_loc(
  object = NULL,
  response = NULL,
  weights = NULL,
  ETA = NULL,
  bayes_para = NULL,
  verbose = FALSE,
  core,
  ...
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
