% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_execute_para.R
\name{model_execute}
\alias{model_execute}
\title{Title}
\usage{
model_execute(
  pheno_data = NULL,
  pheno_data_train = NULL,
  pheno_data_test = NULL,
  geno_data = NULL,
  omic1_data = NULL,
  omic2_data = NULL,
  omic3_data = NULL,
  omics_data_label = list(omic1_data = NULL, omic2_data = NULL, omic3_data = NULL),
  gmatrix = NULL,
  gkernel = NULL,
  pedigree_matrix = NULL,
  omic1_kernel = NULL,
  omic2_kernel = NULL,
  omic3_kernel = NULL,
  omics_kernel_label = list(omic1_kernel = NULL, omic2_kernel = NULL, omic3_kernel =
    NULL),
  train_geno_data = NULL,
  train_omic1_data = NULL,
  train_omic2_data = NULL,
  train_omic3_data = NULL,
  train_omics_label = list(train_omic1_data = NULL, train_omic2_data = NULL,
    train_omic3_data = NULL),
  test_geno_data = NULL,
  test_omic1_data = NULL,
  test_omic2_data = NULL,
  test_omic3_data = NULL,
  test_omics_label = list(test_omic1_data = NULL, test_omic2_data = NULL, test_omic3_data
    = NULL),
  coefficient_1 = NULL,
  coefficient_2 = NULL,
  coefficient_3 = NULL,
  coefficient_4 = NULL,
  train_set = NULL,
  test_set = NULL,
  gmatrix_method = NULL,
  kernel_method = NULL,
  response = NULL,
  gen_name = NULL,
  cova = NULL,
  fixed = NULL,
  random = NULL,
  heter_resid = FALSE,
  heter_groups = NULL,
  VarCov_str = NULL,
  weights = NULL,
  nIter = NULL,
  burnIn = NULL,
  thin = NULL,
  GS_model = c("GBLUP", "GBLUP_BRR", "RKHS", "BRR", "BayesA", "BayesB", "BayesC", "BL",
    "Xgboost", "RandomForest", "PartialLeastSquare", "SupportVectorMachine",
    "K-NearestNeighbors", "Lasso", "Ridge_Regression"),
  eval_metrics = c("Accuracy", "Mean_Squared_Error", "Bias", "Root_Mean_Squared_Error",
    "Relative_Squared_Error", "Mean_Absolute_Error", "Mean_Absolute_Percent_Error"),
  para_tunning = FALSE,
  fixed_term_model_bayesian = "FIXED",
  rand_term_model_bayesian = NULL,
  core = NULL,
  engine = NULL,
  center = TRUE,
  scaled = TRUE,
  workspace = 1e+08,
  pworkspace = 1e+06,
  maxit = 50,
  bending = TRUE,
  bend_value = 0.01,
  blending = FALSE,
  blending_value = 0.02,
  high_diag_cut_off = 1.2,
  low_diag_cut_off = 0.8,
  duplicate_cut_off = 0.95,
  rcn_cutoff = 1e-12,
  optimize_diagonal = FALSE,
  optimize_duplicate = FALSE,
  message = TRUE,
  system_database = FALSE,
  ...
)
}
\arguments{
\item{pheno_data}{phenotypic data, which can be contain both training and testing set. NA is allowed. Dataframe or matrix is allowed}

\item{pheno_data_train}{phenotypic data for training set. NA not allowed. Dataframe or matrix is allowed}

\item{pheno_data_test}{phenotypic data for the testing set. NA is allowed. Dataframe or matrix is allowed}

\item{geno_data}{Genomic/SNP/Marker data NA is allowed but not expected.
numeric 0, 1, 2 (where 0 is minor allele, 1 is hetero and 2 is major allele)
and -1, 0, 1 is also allowed (where -1 is minor allele, 0 is hetero and 1 is major allele).
 Dataframe or matrix is allowed
 We allowed up to 4 different omics data for model fit}

\item{omic1_data}{Omic data (transcriptomic, metabolic, proteomic, environment etc) NA is allowed but not expected. Dataframe or matrix is allowed}

\item{omic2_data}{Similar to Omic1_data}

\item{omic3_data}{Similar to Omic1_data}

\item{omics_data_label}{lable/name of the omics data}

\item{gmatrix}{Genomic relationship matrix, NA not allowed. Dataframe or matrix is allowed}

\item{gkernel}{relationship matrix using different kernel methods}

\item{omic1_kernel}{relationship matrix using different kernel methods}

\item{omic2_kernel}{relationship matrix using different kernel methods}

\item{omic3_kernel}{relationship matrix using different kernel methods}

\item{omics_kernel_label}{lable/name for the omics_kernel if any}

\item{train_geno_data}{Genomic data for training set if geno_data is not provided by the user. Dataframe or matrix is allowed}

\item{train_omic1_data}{Omic data for training set}

\item{train_omic2_data}{Omic data for training set}

\item{train_omic3_data}{Omic data for training set}

\item{test_geno_data}{Genomic data for testing set if geno_data is not provided or not included in the geno_data by the user.
In that scenario geno_data is considered training set and training_geno_data is not provided by the user. Dataframe or matrix is allowed}

\item{test_omic1_data}{Omic data for testing set if omic1_data is not provided or not included in the omic1_data by the user.
In that scenario omic1_data is considered training set and training_geno_data is not provided by the user. Dataframe or matrix is allowed}

\item{test_omic2_data}{same as test_omic1_data}

\item{test_omic3_data}{same as test_omic1_data}

\item{coefficient_1}{coefficient for the training set using either genomic or any omics data.
We allowed up to 4 omics data for model fit}

\item{train_set}{Dataframe with column name of the individual in the training set. This is useful maining
for purpose of cross-validation exercise.}

\item{test_set}{Dataframe with column name of the individual in the testing set. Not required
if pheno_data contain individuals (testing set) with no phenotypic record as NA.}

\item{gmatrix_method}{two methods are currently available to calculate the genomic relationship matrix
Yang and Van-raden}

\item{kernel_method}{kernel methods to calculate relationship matrix for the
different omics. Currently available are Gaussian kernel, exponential kernel
Polynomia kernel (order 2, 3, 4) and linear}

\item{response}{trait(s) of interest to the user}

\item{gen_name}{Column name containing individuals/genotypes}

\item{cova}{covariate if any.Its epected in formula i.e cova  = ~ Rain + Temp}

\item{fixed}{fixed terms. Its expected in formula i.e fixed = ~ name + Env}

\item{random}{random terms. Its expected in formula i.r random = ~ name + Env}

\item{heter_resid}{True or False if user want heterogeneous residual variance or not}

\item{heter_groups}{Column name for Environment or location}

\item{VarCov_str}{user defined variance-covariance structure}

\item{weights}{weight for the response variable. Only dataframe}

\item{nIter}{number of iteration for Bayesian models}

\item{burnIn}{number of burnin  for Bayesian models}

\item{thin}{number of thinning for Bayesian models}

\item{GS_model}{GS-model for fit. The following are available
BayesA, BayesB, BayesC, Baysian Ridge Regression (BRR). These models only work with
M-matrix(genomic data and omic data) in single location.
Bayesian reproducing kernel Hilbert spaces regressions (RKHS),
Bayesian Genomic Best linear unbias estimate (BGBLUP). Both RKHS and BGBLUP can
fit both single and multiple location using reaction norm.
Genomic Best linear unbias estimate using asreml-R package with different
variance structure such as (FA, RR, US, CORGH, CORGV, CORH, CORV)
for multi-location/environment.
Machine learning models include:
Extreme Gradiant Boosting, Random Forest, KNN, Lasso, Ridge Regression,
Partial Least Square, Support Vector Machine. All the machine learning only work
in single location.}

\item{fixed_term_model_bayesian}{model for the fixed term which is always fixed}

\item{rand_term_model_bayesian}{model for the random terms which can be any of the above mentioned model}

\item{core}{number of ram for paralllel job}

\item{engine}{if user has asreml}

\item{center}{if X-variables should be standardized}

\item{workspace}{allocate memory for asreml model fit}

\item{pworkspace}{allocate memory for predict function in asreml}

\item{maxit}{number of iteration for asreml}

\item{bending}{this is important when the relationship matrix is not positive definitive. It fix it for the user. it has be TRUE}

\item{message}{if message/warning should be displayed}

\item{system_database}{this dictate if the output will be created in a folder or as list
the default is FALSE. Thus output will be folder.}

\item{...}{}
}
\description{
Title
}
\examples{
#######
1. It check if the matrix is square matrix/symmetry, if not we fix it for the user
2. It check if the matrix is positive definite, if not we fix it.
3. It check for NA. If present the engine will stop further analysis.
}
