% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omic_to_model.R
\name{omic_to_model}
\alias{omic_to_model}
\title{Title}
\usage{
omic_to_model(
  omic_data = NULL,
  train_omic_data = NULL,
  test_omic_data = NULL,
  message = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
