% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{phenotype}
\alias{phenotype}
\title{Prepares phenotypic data}
\usage{
phenotype(
  object = NULL,
  to.numeric = NULL,
  to.factor = NULL,
  noNAon = NULL,
  message = TRUE
)
}
\description{
Prepares phenotypic data
}
\examples{

# Messing with the data.
nassau <- asreml::nassau
nassau$weight <- 1 ; nassau$weight[1] <- NA
nassau$Rep <- as.numeric(nassau$Rep)

nassau. <- nassau |>
  phenotype(
    to.factor = c("Rep"),
    to.numeric = c("IncBlock"),
    noNAon = c("ht6", "weight"))

nassau. |> str()

nassau. |> class()

}
