% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_precheck.R
\name{phenotype_precheck}
\alias{phenotype_precheck}
\title{Title}
\usage{
phenotype_precheck(
  pheno_data = NULL,
  gen_name = NULL,
  response = NULL,
  heter_groups = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
This function address the following objectives:
1. Check if the phenotypic data file contain records
2. check if the phenotypic data is in data.frame and if not we fit it for the user
3. Check if the gen_name and response names provided by the user match the column
   names in the phenotypic data
4. Check that the gen_name did not has NA as part of the genotypes ID
5. Check that the response variable is numeric if not convert it to numeric
7. Check for NA is not test here for response variables because testing set might have NA
   thus, NA is expected.
6. If all these checks were fulfilled assign class(pheno_data) <- c("data.frame", "phenotype")
and define attribute as attr(pheno_data, "cleared") <- "pass"
}
\details{
7. This output will be input for phenotype_to_model function
}
