% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_geno_omic_kernel_gmatrix_data.R
\name{process_geno_data}
\alias{process_geno_data}
\title{Title}
\usage{
process_geno_data(
  geno_data = NULL,
  train_geno_data = NULL,
  test_geno_data = NULL,
  test_set = NULL,
  train_set = NULL,
  gen_name = NULL,
  kernel_method = NULL,
  gmatrix_method = NULL,
  scale = NULL,
  map_data = NULL,
  maf_threshold = NULL,
  het_threshold = NULL,
  ind_call_rate_threshold = NULL,
  snp_call_rate_threshold = NULL,
  impute = NULL,
  qc_filtering = NULL,
  message = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
