% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_fix_check.R
\name{rand_fix_check}
\alias{rand_fix_check}
\title{Title
This function do the following:}
\usage{
rand_fix_check(pheno_data = NULL, rand_fix_term = NULL, ...)
}
\arguments{
\item{pheno_data}{phenotypic data}

\item{rand_fix_term}{this can be random or fixed term and expected to be class formula}
}
\description{
1. Check if the random and fixed term are in formula. If not it will stop
2. Check if the ranom and fixed term are in class factor in the pheno_data,
   if not we fix it for the user
3. The output will be class formula and attribute pass
}
