% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesians_preprocess.R
\name{random_term_model}
\alias{random_term_model}
\title{Title}
\usage{
random_term_model(
  rand_terms = NULL,
  GS_model = NULL,
  gen_name = NULL,
  rand_terms_model_bayesian = NULL,
  message = TRUE
)
}
\arguments{
\item{rand_terms}{all the random terms in the formula. This is an output from random_terms function}

\item{GS_model}{model defined by the user. It is expected this will be for the genotype effect estimate}

\item{gen_name}{Column name containing the individuals/genotype in the phenotypic data file}

\item{rand_terms_model_bayesian}{model for all the random terms including the genotype effect.
the total number of the model must be equal the number of random term.}

\item{message}{If the user want to print out the message/warning info or not}
}
\description{
Title
}
