% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cals_sep_pev_rel_gblup.R
\name{sep_pev_rel_gblup}
\alias{sep_pev_rel_gblup}
\title{Title}
\usage{
sep_pev_rel_gblup(
  geno_object = NULL,
  X = NULL,
  Z = NULL,
  va = NULL,
  ve = NULL,
  h2 = NULL,
  lambda = NULL
)
}
\arguments{
\item{lambda}{}
}
\description{
Title
}
