% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics_AI.R
\name{summary_statistics_AI}
\alias{summary_statistics_AI}
\title{Title}
\usage{
summary_statistics_AI(
  mod = NULL,
  pheno_object = NULL,
  response = NULL,
  test_set = NULL,
  geno_omic_object = NULL,
  eval_metrics = c("Accuracy", "Mean_Squared_Error", "Bias", "Root_Mean_Squared_Error",
    "Relative_Squared_Error", "Mean_Absolute_Error", "Mean_Absolute_Percent_Error"),
  GS_model = NULL,
  ...
)
}
\arguments{
\item{eval_metrics}{}
}
\description{
Title
}
