% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_recode_and_qc_vcf.R
\name{vcf_qc_recode}
\alias{vcf_qc_recode}
\title{Title}
\usage{
vcf_qc_recode(
  vcf_file_name = NULL,
  vcf_file_path = NULL,
  vcf_file = NULL,
  maf_threshold = 0.01,
  het_threshold = 0.2,
  ind_call_rate_threshold = 0.2,
  snp_call_rate_threshold = 0.3,
  impute = FALSE,
  recode_format = "0,1,2",
  out_put_map = TRUE,
  message = TRUE
)
}
\arguments{
\item{message}{}
}
\description{
Title
}
